detect current configuration file

+++ scripts/acg.sh
@@ -186,6 +186,15 @@ do_install_acg() {
       echo "Cannot satisfy the requirements. Please check your internet and your apk repositories. Make sure to enable the community repo."
     fi
   fi
+  # check configure file
+  if [ -f "${acg_path}/files/acg-cfg" ]; then
+    if (whiptail --title "Alpine Clash Gateway" --yesno "You already have an ACG configuration file in your system. Continue installation may loss all your original settings. Do you want to continue?" 10 60) then
+      [ -f "${acg_path}/files/acg-cfg-sample" ] && cp "${acg_path}/files/acg-cfg-sample" "${acg_path}/files/acg-cfg"
+    else
+      exit 1
+    fi
+  fi
+
   if (whiptail --title "Alpine Clash Gateway" --yesno "This script will install Alpine Clash Gateway (ACG). It will turn your Alpine Linux host into a Clash gateway. Do you want to continue?" 10 60) then
     echo "Install all dependent packages"
     apk add --update nftables iproute2 udev curl jq
